benchmark/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-customer-service-benchmark"
version = "0.1.0"
description = "Benchmarking LLM capabilities in customer service contexts"
authors = [
    {name = "Artur Pietrzak", email = "artur.pietrzak@example.com"},
]
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.9.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "google-generativeai>=0.3.0",
    "groq>=0.4.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.0.0",
    "pydantic-settings>=2.0.0",
    "jsonschema>=4.17.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "mypy",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100